chore: update dependencies & refactor path classes
