+++ package.json
@@ -13,7 +13,7 @@
   "sideEffects": false,
   "scripts": {
     "start": "webpack serve --mode development --port 3000",
-    "build": "tsc --build ./tsconfig.build.json && webpack --mode production",
+    "build": "tsc --build --clean ./tsconfig.build.json && webpack --mode production",
     "lint": "eslint src/*.ts --fix",
     "test": "npx playwright test"
   },
@@ -25,20 +25,25 @@
   "homepage": "https://github.com/tarkant/svg-dom-arrows#readme",
   "devDependencies": {
     "@playwright/test": "^1.31.1",
-    "@typescript-eslint/eslint-plugin": "^4.18.0",
-    "@typescript-eslint/parser": "^4.18.0",
+    "@types/expect-puppeteer": "^4.4.7",
+    "@types/puppeteer": "^5.4.4",
+    "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "@typescript-eslint/eslint-plugin-tslint": "^4.18.0",
+    "@typescript-eslint/parser": "^4.33.0",
+    "concurrently": "^8.0.1",
     "copy-webpack-plugin": "^8.0.0",
     "css-loader": "^5.1.3",
-    "eslint": "^7.22.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.8.0",
     "html-webpack-plugin": "^5.3.1",
     "sass": "^1.32.8",
     "sass-loader": "^11.0.1",
     "style-loader": "^2.0.0",
     "ts-loader": "^8.0.18",
-    "typescript": "^4.2.4",
+    "typescript": "^4.9.5",
     "url-loader": "^4.1.1",
     "webpack": "^5.65.0",
     "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.7.1","concurrently": "^8.0.1"
+    "webpack-dev-server": "^4.7.1"
   }
 }
