+++ src/paths/index.ts
@@ -1,3 +1,4 @@
+export * from './Path';
 export * from './ArcPath';
 export * from './CurvyPath';
 export * from './LinePath';
